refactor(api): move UserCantBeCreatedError
