+++ api/lib/application/error-manager.js
@@ -147,9 +147,6 @@ function _mapToHttpError(error) {
   if (error instanceof DomainErrors.AlreadySharedCampaignParticipationError) {
     return new HttpErrors.PreconditionFailedError(error.message);
   }
-  if (error instanceof DomainErrors.UserCantBeCreatedError) {
-    return new HttpErrors.UnauthorizedError(error.message);
-  }
   if (error instanceof DomainErrors.MembershipCreationError) {
     return new HttpErrors.BadRequestError(error.message);
   }
